+++ lib/ui/home/tabs/tasbeh/tasbeh_tab.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../onboardingScreenAndSb7a-salmaMourad/sb7aPages/Sb7aPageScreen.dart';
+import '../../../../sb7a/sb7aPages/Sb7aPageScreen.dart';
 
 class TasbehTab extends StatelessWidget {
   const TasbehTab({super.key});
